drivers: dma-contiguous: fix dts overriding setting
The check on cmdline was moved below the other checks so the trace
doesn't appear on a memory block that isn't meant for cma anyway.

This overriding can only apply to the "linux,cma-default" node, which is
really what the cma= parameter is about (being the linux default cma
region).

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>